refactor(cydiffer): move applyDiff from cyUpdateHelpers to cyDiffer
